properly separate preprocessing from pca analysis
